requirements/base.txt
# Core
Django==3.2

# Admin
django-jet==1.0.8
django-js-asset==1.2.2
django-admin-tools==0.9.0
django-fluent-dashboard==1.0.1

# Configuration
django-split-settings==1.2.0

# Asset
django-bootstrap4==2.3.1
django-markdown-deux==1.0.5
Markdown==3.3.3
markdown2==2.3.10
Pillow==9.3.0

# Forms
django-ckeditor==6.0.0
django-crispy-forms==1.10.0

# Envs
django-decouple==2.1
python-decouple==3.3

# REST
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0

# Filter
django-filter==2.4.0

# Utils
django-utils-six==2.0
